Migrated helloapp from python venv to a docker image
Created a simple docker image, using the official v3.12 slim python image as the base image
As we are creating a new isolated container environment, there is no need to maintain a python virtual environment
We have created an 'entrypoint.sh' bash script, to be executed at runtime. This script accepts a single parameter referencing the script to run
The command argument options are 'test' and 'run' (default). 'test.sh' and 'run.sh' also work.

  - 'test.sh' - runs the python unittest module on ./helloapp/test.py
  - 'run.sh' - runs the helloapp application via gunicorn exposing the Flask app on port 8080

To assist with local development, a 'Makefile' has been created with the following available targets:

  - 'build' - Runs a docker build tagging the image as 'helloapp:0.0.1' unless overridden with 'image-name' and/or 'image-tag'
  - 'test'  - Runs the `test` command on the docker image
  - 'run'   - Runs the `run` command on the docker image
  - 'shell' - Opens a bash shell in the container for development or debugging purposes

'build', 'run' and 'shell' targets all trigger a build first unless overridden with 'build=false'

The Dockerfile to build can be overridden with the argument 'dockerfile'
The port 'run' exposes can be overriden with the argument 'port'

Resolves: Task #1